Make it so that only an Organizer can register, update and delete Meetups.

+++ Meetekat.WebApi/Features/Meetups/DeleteMeetup/DeleteMeetupFeature.cs
@@ -2,8 +2,10 @@
 
 using System;
 using System.Linq;
+using Meetekat.WebApi.Entities.Users;
 using Meetekat.WebApi.Persistence;
 using Meetekat.WebApi.Seedwork.Features;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -17,8 +19,10 @@ public class DeleteMeetupFeature : FeatureBase
 
     [Tags(ApiSections.Meetups)]
     [HttpDelete("/api/meetups/{meetupId:guid}")]
+    [Authorize(Roles = nameof(Organizer))]
     [SwaggerOperation("Delete a Meetup with matching ID.")]
     [SwaggerResponse(StatusCodes.Status204NoContent, "A Meetup with the specified ID was deleted successfully.")]
+    [SwaggerResponse(StatusCodes.Status403Forbidden, "Only a Meetup's direct Organizer can delete the Meetup.")]
     [SwaggerResponse(StatusCodes.Status404NotFound, "A Meetup with the specified ID doesn't exist.")]
     public IActionResult DeleteMeetup([FromRoute] Guid meetupId)
     {
@@ -26,6 +30,18 @@ public class DeleteMeetupFeature : FeatureBase
         if (meetup is null)
         {
             return NotFound();
+        }
+        
+        var organizerExists = context.Organizers.Any(organizer => organizer.Id == Caller.UserId);
+        if (!organizerExists)
+        {
+            // Can happen if deleted Organizer tries to delete a Meetup (if the Access Token hasn't yet expired).
+            return Unauthorized();
+        }
+        if (meetup.OrganizerId != Caller.UserId)
+        {
+            // Only a Meetup's direct Organizer can delete the Meetup.
+            return Forbidden();
         }
 
         context.Meetups.Remove(meetup);
